Data.ByteArray.Base64String

Documentation

newtype Base64String

Represents a Hex string. Guarantees that all characters it contains are valid hex characters.

Constructors

Base64String

Fields

unBase64String :: ByteString

base64String :: ByteArray ba => ba -> Either String Base64String

Smart constructor that works with any mixed casing of characters: `Base64String AA == Base64String "0xAa" == Base64String "0xaA" == Base64String "aa"`

fromBytes :: ByteArrayAccess ba => ba -> Base64String

Reads a raw bytes and converts to hex representation.

toBytes :: ByteArray ba => Base64String -> ba

Access to the raw bytes of Base64String.

toText :: Base64String -> Text

Access to a Text representation of the Base64String

format :: Base64String -> Text

Access to a Text representation of the Base64String