Data.ByteArray.HexString

Documentation

newtype HexString

Represents a Hex string. Guarantees that all characters it contains are valid hex characters.

Constructors

HexString

Fields

unHexString :: ByteString

hexString :: ByteArray ba => ba -> Either String HexString

Smart constructor which trims '0x' and validates length is even. Works with any mixed casing of characters: `hexString "0xAA" == hexString "0xAa" == hexString "0xaA" == hexString "0xaa"`

fromBytes :: ByteArrayAccess ba => ba -> HexString

Reads a raw bytes and converts to hex representation.

toBytes :: ByteArray ba => HexString -> ba

Access to the raw bytes of HexString.

toText :: HexString -> Text

Access to a Text representation of the HexString

format :: HexString -> Text

Access to a Text representation of the HexString