Network.ABCI.Types.Messages.FieldTypes

Documentation

newtype WrappedVal a

Constructors

WrappedVal

Fields

unWrappedVal :: a

data Timestamp

Constructors

Timestamp DiffTime

mkTimestamp :: DiffTime -> Timestamp

parseDiffTimeOrError :: String -> DiffTime

data BlockParams

Constructors

BlockParams

Fields

blockParamsMaxBytes :: WrappedVal Int64

Max size of a block, in bytes.

blockParamsMaxGas :: WrappedVal Int64

Max sum of GasWanted in a proposed block.

data EvidenceParams

Constructors

EvidenceParams

Fields

evidenceParamsMaxAge :: WrappedVal Int64

Max age of evidence, in blocks.

data ValidatorParams

Constructors

ValidatorParams

Fields

validatorParamsPubKeyTypes :: [Text]

List of accepted pubkey types

data ConsensusParams

Constructors

ConsensusParams

Fields

consensusParamsBlockSize :: Maybe BlockParams

consensusParamsEvidence :: Maybe EvidenceParams

Parameters limiting the validity of evidence of byzantine behaviour.

consensusParamsValidator :: Maybe ValidatorParams

Parameters limitng the types of pubkeys validators can use.

data PubKey

Constructors

PubKey

Fields

pubKeyType :: Text

Type of the public key.

pubKeyData :: Base64String

Public key data.

data ValidatorUpdate

Constructors

ValidatorUpdate

Fields

validatorUpdatePubKey :: Maybe PubKey

Public key of the validator

validatorUpdatePower :: WrappedVal Int64

Voting power of the validator

data Validator

Constructors

Validator

Fields

validatorAddress :: HexString

Address of the validator (hash of the public key)

validatorPower :: WrappedVal Int64

Voting power of the validator

data VoteInfo

Constructors

VoteInfo

Fields

voteInfoValidator :: Maybe Validator

A validator

voteInfoSignedLastBlock :: Bool

Indicates whether or not the validator signed the last block

data LastCommitInfo

Constructors

LastCommitInfo

Fields

lastCommitInfoRound :: WrappedVal Int32

Commit round.

lastCommitInfoVotes :: [VoteInfo]

List of validators addresses in the last validator set with their voting power and whether or not they signed a vote.

data PartSetHeader

Constructors

PartSetHeader

Fields

partSetHeaderTotal :: WrappedVal Int32

total number of pieces in a PartSet

partSetHeaderHash :: HexString

Merkle root hash of those pieces

data BlockID

Constructors

BlockID

Fields

blockIDHash :: HexString

Hash of the block header

blockIDPartsHeader :: Maybe PartSetHeader

PartSetHeader (used internally, for clients this is basically opaque).

data Version

Constructors

Version

Fields

versionBlock :: WrappedVal Word64

Protocol version of the blockchain data structures.

versionApp :: WrappedVal Word64

Protocol version of the application.

data Header

Constructors

Header

Fields

headerVersion :: Maybe Version

Version of the blockchain and the application

headerChainId :: Text

ID of the blockchain

headerHeight :: WrappedVal Int64

Height of the block in the chain

headerTime :: Maybe Timestamp

Time of the previous block

headerNumTxs :: WrappedVal Int64

Number of transactions in the block

headerTotalTxs :: WrappedVal Int64

Total number of transactions in the blockchain until now

headerLastBlockId :: Maybe BlockID

Hash of the previous (parent) block

headerLastCommitHash :: HexString

Hash of the previous block's commit

headerDataHash :: HexString

Hash of the validator set for this block

headerValidatorsHash :: HexString

Hash of the validator set for the next block

headerNextValidatorsHash :: HexString

Hash of the consensus parameters for this block

headerConsensusHash :: HexString

Hash of the consensus parameters for this block

headerAppHash :: HexString

Data returned by the last call to Commit

headerLastResultsHash :: HexString

Hash of the ABCI results returned by the last block

headerEvidenceHash :: HexString

Hash of the evidence included in this block

headerProposerAddress :: HexString

Original proposer for the block

data Evidence

Constructors

Evidence

Fields

evidenceType :: Text

Type of the evidence.

evidenceValidator :: Maybe Validator

The offending validator

evidenceHeight :: WrappedVal Int64

Height when the offense was committed

evidenceTime :: Maybe Timestamp

Time of the block at height Height.

evidenceTotalVotingPower :: WrappedVal Int64

Total voting power of the validator set at height Height

data KVPair

Constructors

KVPair

Fields

kVPairKey :: Base64String

key

kVPairValue :: Base64String

value

data Proof

Constructors

Proof

Fields

proofOps :: [ProofOp]

List of chained Merkle proofs, of possibly different types

data ProofOp

Constructors

ProofOp

Fields

proofOpType :: Text

Type of Merkle proof and how it's encoded.

proofOpKey :: Base64String

Key in the Merkle tree that this proof is for.

proofOpData :: Base64String

Encoded Merkle proof for the key.

data Event

Constructors

Event

Fields

eventType :: Text

Type of Event

eventAttributes :: [KVPair]

Event attributes