Network.ABCI.Types.Messages.Request

Documentation

data Echo

Constructors

Echo

Fields

echoMessage :: Text

A string to echo back

_echoMessage :: Iso' Echo Text

data Flush

Constructors

Flush

data Info

Constructors

Info

Fields

infoVersion :: Text

The Tendermint software semantic version

infoBlockVersion :: Word64

The Tendermint Block Protocol version

infoP2pVersion :: Word64

The Tendermint P2P Protocol version

_infoVersion :: Lens' Info Text

_infoP2pVersion :: Lens' Info Word64

_infoBlockVersion :: Lens' Info Word64

data SetOption

Constructors

SetOption

Fields

setOptionKey :: Text

Key to set

setOptionValue :: Text

Value to set for key

_setOptionValue :: Lens' SetOption Text

_setOptionKey :: Lens' SetOption Text

data InitChain

Constructors

InitChain

Fields

initChainTime :: Maybe Timestamp

Genesis time

initChainChainId :: Text

ID of the blockchain.

initChainConsensusParams :: Maybe ConsensusParams

Initial consensus-critical parameters.

initChainValidators :: [ValidatorUpdate]

Initial genesis validators.

initChainAppState :: Base64String

Serialized initial application state. Amino-encoded JSON bytes.

_initChainValidators :: Lens' InitChain [ValidatorUpdate]

_initChainTime :: Lens' InitChain (Maybe Timestamp)

_initChainConsensusParams :: Lens' InitChain (Maybe ConsensusParams)

_initChainChainId :: Lens' InitChain Text

_initChainAppState :: Lens' InitChain Base64String

data Query

Constructors

Query

Fields

queryData :: Base64String

Raw query bytes. Can be used with or in lieu of Path.

queryPath :: Text

Path of request, like an HTTP GET path. Can be used with or in liue of Data.

queryHeight :: WrappedVal Int64

The block height for which you want the query

queryProve :: Bool

Return Merkle proof with response if possible

_queryProve :: Lens' Query Bool

_queryPath :: Lens' Query Text

_queryHeight :: Lens' Query (WrappedVal Int64)

_queryData :: Lens' Query Base64String

data BeginBlock

Constructors

BeginBlock

Fields

beginBlockHash :: HexString

The block's hash. This can be derived from the block header.

beginBlockHeader :: Maybe Header

The block header.

beginBlockLastCommitInfo :: Maybe LastCommitInfo

Info about the last commit, including the round, and the list of validators and which ones signed the last block.

beginBlockByzantineValidators :: [Evidence]

List of evidence of validators that acted maliciously.

_beginBlockLastCommitInfo :: Lens' BeginBlock (Maybe LastCommitInfo)

_beginBlockHeader :: Lens' BeginBlock (Maybe Header)

_beginBlockHash :: Lens' BeginBlock HexString

_beginBlockByzantineValidators :: Lens' BeginBlock [Evidence]

data CheckTx

Constructors

CheckTx

Fields

checkTxTx :: Base64String

The request transaction bytes

_checkTxTx :: Iso' CheckTx Base64String

data DeliverTx

Constructors

DeliverTx

Fields

deliverTxTx :: Base64String

The request transaction bytes.

_deliverTxTx :: Iso' DeliverTx Base64String

data EndBlock

Constructors

EndBlock

Fields

endBlockHeight :: WrappedVal Int64

Height of the block just executed.

_endBlockHeight :: Iso' EndBlock (WrappedVal Int64)

data Commit

Constructors

Commit