Network.ABCI.Types.Messages.Response

Documentation

data Echo

Constructors

Echo

Fields

echoMessage :: Text

The input string

_echoMessage :: Iso' Echo Text

data Flush

Constructors

Flush

data Info

Constructors

Info

Fields

infoData :: Text

Some arbitrary information

infoVersion :: Text

The application software semantic version

infoAppVersion :: WrappedVal Word64

The application protocol version

infoLastBlockHeight :: WrappedVal Int64

Latest block for which the app has called Commit

infoLastBlockAppHash :: HexString

Latest result of Commit

_infoVersion :: Lens' Info Text

_infoLastBlockHeight :: Lens' Info (WrappedVal Int64)

_infoLastBlockAppHash :: Lens' Info HexString

_infoData :: Lens' Info Text

_infoAppVersion :: Lens' Info (WrappedVal Word64)

data SetOption

Constructors

SetOption

Fields

setOptionCode :: Word32

Response code

setOptionLog :: Text

The output of the application's logger. May be non-deterministic.

setOptionInfo :: Text

Additional information. May be non-deterministic.

_setOptionLog :: Lens' SetOption Text

_setOptionInfo :: Lens' SetOption Text

_setOptionCode :: Lens' SetOption Word32

data InitChain

Constructors

InitChain

Fields

initChainConsensusParams :: Maybe ConsensusParams

Initial consensus-critical parameters.

initChainValidators :: [ValidatorUpdate]

Initial validator set (if non empty).

_initChainValidators :: Lens' InitChain [ValidatorUpdate]

_initChainConsensusParams :: Lens' InitChain (Maybe ConsensusParams)

data Query

Constructors

Query

Fields

queryCode :: Word32

Response code.

queryLog :: Text

The output of the application's logger. May be non-deterministic.

queryInfo :: Text

Additional information. May be non-deterministic.

queryIndex :: WrappedVal Int64

The index of the key in the tree.

queryKey :: Base64String

The key of the matching data.

queryValue :: Base64String

The value of the matching data.

queryProof :: Maybe Proof

Serialized proof for the value data, if requested, to be verified against the AppHash for the given Height.

queryHeight :: WrappedVal Int64

The block height from which data was derived.

queryCodespace :: Text

Namespace for the Code.

_queryValue :: Lens' Query Base64String

_queryProof :: Lens' Query (Maybe Proof)

_queryLog :: Lens' Query Text

_queryKey :: Lens' Query Base64String

_queryInfo :: Lens' Query Text

_queryIndex :: Lens' Query (WrappedVal Int64)

_queryHeight :: Lens' Query (WrappedVal Int64)

_queryCodespace :: Lens' Query Text

_queryCode :: Lens' Query Word32

data BeginBlock

Constructors

BeginBlock

Fields

beginBlockEvents :: [Event]

Beginning block events

_beginBlockEvents :: Iso' BeginBlock [Event]

data CheckTx

Constructors

CheckTx

Fields

checkTxCode :: Word32

Response code

checkTxData :: Base64String

Result bytes, if any.

checkTxLog :: Text

The output of the application's logger.

checkTxInfo :: Text

Additional information.

checkTxGasWanted :: WrappedVal Int64

Amount of gas requested for transaction.

checkTxGasUsed :: WrappedVal Int64

Amount of gas consumed by transaction.

checkTxEvents :: [Event]

Events

checkTxCodespace :: Text

Namespace for the Code.

_checkTxLog :: Lens' CheckTx Text

_checkTxInfo :: Lens' CheckTx Text

_checkTxGasWanted :: Lens' CheckTx (WrappedVal Int64)

_checkTxGasUsed :: Lens' CheckTx (WrappedVal Int64)

_checkTxEvents :: Lens' CheckTx [Event]

_checkTxData :: Lens' CheckTx Base64String

_checkTxCodespace :: Lens' CheckTx Text

_checkTxCode :: Lens' CheckTx Word32

data DeliverTx

Constructors

DeliverTx

Fields

deliverTxCode :: Word32

Response code.

deliverTxData :: Base64String

Result bytes, if any.

deliverTxLog :: Text

The output of the application's logger. May be non-deterministic.

deliverTxInfo :: Text

Additional information.

deliverTxGasWanted :: WrappedVal Int64

Amount of gas requested for transaction.

deliverTxGasUsed :: WrappedVal Int64

Amount of gas consumed by transaction.

deliverTxEvents :: [Event]

Events

deliverTxCodespace :: Text

Namespace for the Code.

_deliverTxLog :: Lens' DeliverTx Text

_deliverTxInfo :: Lens' DeliverTx Text

_deliverTxGasWanted :: Lens' DeliverTx (WrappedVal Int64)

_deliverTxGasUsed :: Lens' DeliverTx (WrappedVal Int64)

_deliverTxEvents :: Lens' DeliverTx [Event]

_deliverTxData :: Lens' DeliverTx Base64String

_deliverTxCodespace :: Lens' DeliverTx Text

_deliverTxCode :: Lens' DeliverTx Word32

data EndBlock

Constructors

EndBlock

Fields

endBlockValidatorUpdates :: [ValidatorUpdate]

Changes to validator set (set voting power to 0 to remove).

endBlockConsensusParamUpdates :: Maybe ConsensusParams

Changes to consensus-critical time, size, and other parameters.

endBlockEvents :: [Event]

Events

_endBlockValidatorUpdates :: Lens' EndBlock [ValidatorUpdate]

_endBlockEvents :: Lens' EndBlock [Event]

_endBlockConsensusParamUpdates :: Lens' EndBlock (Maybe ConsensusParams)

data Commit

Constructors

Commit

Fields

commitData :: Base64String

The Merkle root hash of the application state

_commitData :: Iso' Commit Base64String

data Exception

Constructors

Exception

Fields

exceptionError :: Text

_exceptionError :: Iso' Exception Text